+++ CHANGELOG.rst
@@ -1,7 +1,11 @@
 1.3.6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- directly provide base request interfaces instead of dynamically applying them
+  [vangheem]
+
+- Provide iter_data method on cloud file manager
+  [vangheem]
 
 
 1.3.5 (2017-08-08)